write platform specific code with different syntax
